pyproject.toml
[tool.poetry]
name = "erc"
version = "0.1.0"
description = ""
authors = ["ren <denis30510@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
ipython = "^8.18.1"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
